+++ src/gltf/writer.ts
@@ -2,11 +2,11 @@ import * as path from 'path';
 import * as fse from 'fs-extra';
 
 import * as gltf from './schema';
-import { isUndefined } from 'util';
+import { isUndefined, isNullOrUndefined } from 'util';
 import { IMaterial, IFragment, IMesh, ILines, IPoints, IMaterialMap } from '../svf/schema';
 import { ISvfContent } from '../svf/reader';
 
-const BufferSizeLimit = 5 << 20;
+const MaxBufferSize = 5 << 20;
 const DefaultMaterial: gltf.MaterialPbrMetallicRoughness = {
     pbrMetallicRoughness: {
         baseColorFactor: [0.25, 0.25, 0.25, 1.0],
@@ -14,6 +14,13 @@ const DefaultMaterial: gltf.MaterialPbrMetallicRoughness = {
         roughnessFactor: 0.5
     }
 };
+
+export interface IWriterOptions {
+    maxBufferSize?: number; /** Approx. size limit (in bytes) of binary buffers with mesh data (5 << 20 by default) */
+    ignoreMeshGeometry?: boolean; /** Don't output mesh geometry */
+    ignoreLineGeometry?: boolean; /** Don't output line geometry */
+    ignorePointGeometry?: boolean; /** Don't output point geometry */
+}
 
 /**
  * Utility class for serializing SVF content to local file system as glTF (2.0).
@@ -24,11 +31,20 @@ export class Writer {
     protected bufferStream: fse.WriteStream | null;
     protected bufferSize: number;
     protected baseDir: string;
+    protected maxBufferSize: number;
+    protected ignoreMeshGeometry: boolean;
+    protected ignoreLineGeometry: boolean;
+    protected ignorePointGeometry: boolean;
 
     /**
      * Initializes the writer.
+     * @param {IWriterOptions} [options={}] Additional writer options.
      */
-    constructor() {
+    constructor(options: IWriterOptions = {}) {
+        this.maxBufferSize = isNullOrUndefined(options.maxBufferSize) ? MaxBufferSize : options.maxBufferSize;
+        this.ignoreMeshGeometry = !!options.ignoreMeshGeometry;
+        this.ignoreLineGeometry = !!options.ignoreLineGeometry;
+        this.ignorePointGeometry = !!options.ignorePointGeometry;
         this.manifest = {
             asset: {
                 version: '2.0',
@@ -79,8 +95,8 @@ export class Writer {
             scene: -1
         };
 
-        const manifestScenes = this.manifest.scenes as gltf.Scene[];
         fse.ensureDirSync(this.baseDir);
+        const manifestScenes = this.manifest.scenes as gltf.Scene[];
         manifestScenes.push(this.writeScene(svf));
         this.manifest.scene = 0;
         const gltfPath = path.join(this.baseDir, 'output.gltf');
@@ -173,10 +189,14 @@ export class Writer {
             primitives: []
         };
 
+        if (this.ignoreMeshGeometry) {
+            return mesh;
+        }
+
         const manifestBuffers = this.manifest.buffers as gltf.Buffer[];
 
         // Prepare new writable stream if needed
-        if (this.bufferStream === null || this.bufferSize > BufferSizeLimit) {
+        if (this.bufferStream === null || this.bufferSize > this.maxBufferSize) {
             if (this.bufferStream) {
                 this.bufferStream.close();
                 this.bufferStream = null;
@@ -332,10 +352,14 @@ export class Writer {
             primitives: []
         };
 
+        if (this.ignoreLineGeometry) {
+            return mesh;
+        }
+
         const manifestBuffers = this.manifest.buffers as gltf.Buffer[];
 
         // Prepare new writable stream if needed
-        if (this.bufferStream === null || this.bufferSize > BufferSizeLimit) {
+        if (this.bufferStream === null || this.bufferSize > this.maxBufferSize) {
             if (this.bufferStream) {
                 this.bufferStream.close();
                 this.bufferStream = null;
@@ -460,10 +484,14 @@ export class Writer {
             primitives: []
         };
 
+        if (this.ignorePointGeometry) {
+            return mesh;
+        }
+
         const manifestBuffers = this.manifest.buffers as gltf.Buffer[];
 
         // Prepare new writable stream if needed
-        if (this.bufferStream === null || this.bufferSize > BufferSizeLimit) {
+        if (this.bufferStream === null || this.bufferSize > this.maxBufferSize) {
             if (this.bufferStream) {
                 this.bufferStream.close();
                 this.bufferStream = null;