+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added
+  - Support for converting both remote and local SVFs using the CLI tool
+  - Support for configuring glTF output (max. size of binary files, ignoring line/point geometries, ...)
+- Changed
+  - Moved to new version of forge-server-utils
+
 ## [0.3.0] - 2019-10-07
 
 - Added