+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2019-10-07
+
+- Added
+  - Access to internal SVF manifest
+- Fixed
+  - Gltf schema now included in build output
+
 ## [0.4.0] - 2019-10-07
 
 - Added